NASA/requirements.txt
beautifulsoup4==4.11.1
blinker==1.8.0
bs4==0.0.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==4.1
Flask==3.0.3
Flask-Cors==4.0.0
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mock==5.1.0
nose==1.3.7
packaging==24.0
pillow==10.3.0
python-dotenv==1.0.1
requests==2.31.0
setupext-janitor==1.0.0
soupsieve==2.5
urllib3==2.2.1
waitress==2.1.2
Werkzeug==3.0.2
